+++ src/Writer/ProjectWriter.php
@@ -49,9 +49,12 @@ class ProjectWriter implements WriterInterface
      */
     use PersistTrait;
 
-    public function save(ObjectInterface $object, PromiseInterface $promise = null): WriterInterface
-    {
-        $this->persist($object, $promise);
+    public function save(
+        ObjectInterface $object,
+        PromiseInterface $promise = null,
+        ?bool $prefereRealDateOnUpdate = null,
+    ): WriterInterface {
+        $this->persist($object, $promise, $prefereRealDateOnUpdate);
 
         return $this;
     }